first version

+++ src/training/dashboard/dashboard.py
@@ -161,10 +161,10 @@ class Dashboard:
             f"pred_{step}": utils.adjust_image_colors(pred_tensor).permute(1, 2, 0),
         }
 
-        torchvision.utils.save_image(images["image_train"], "image.png")
-        torchvision.utils.save_image(images["input_train"], "input.png")
-        torchvision.utils.save_image(images["target_train"], "target.png")
-        torchvision.utils.save_image(images["pred_train"], "pred.png")
+        torchvision.utils.save_image(images["image_train"].type(torch.float32), "image.png")
+        torchvision.utils.save_image(images["input_train"].type(torch.float32), "input.png")
+        torchvision.utils.save_image(images["target_train"].type(torch.float32), "target.png")
+        torchvision.utils.save_image(images["pred_train"].type(torch.float32), "pred.png")
 
         for image_id in images.keys():
             images[image_id] = [wandb.Image(self._tensor_to_pil(images[image_id]))]
